Syntactic issue from previous merge fixed. Explicit cast to long for formatted string for 32-bit compatibility

+++ Visual Scheduler/Category/UIView+BBASubviews.m
@@ -10,7 +10,7 @@
         }
     }
     if (!subview) {
-        NSString *reasonForException = [NSString stringWithFormat:@"No subview found matching tag: %ld", tagOfView];
+        NSString *reasonForException = [NSString stringWithFormat:@"No subview found matching tag: %ld", (long)tagOfView];
         @throw [NSException exceptionWithName:@"View not found" reason:reasonForException userInfo:nil];
     }
     return nil;
